% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/giterateR-package.R
\docType{package}
\name{giterateR-package}
\alias{giterateR}
\alias{giterateR-package}
\title{giterateR: Code Execution Iterater across Git Refs}
\description{
Iteratively execute code across specified Git refs (commits,
    tags, branches). Built on git2r.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/allenbaron/giterateR}
  \item Report bugs at \url{https://github.com/allenbaron/giterateR/issues}
}

}
\keyword{internal}
